+++ tests/irix-g/output/test.c
@@ -1,5 +1,5 @@
 test(void) {
-    (s32) D_410100 = 4;
+    D_410100 = 4;
     func_004000B0();
     return;
     // (possible return value: func_004000B0())
